+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
 - Add `send_config` method to send a single configuration string -- this will automagically handle sending a full
  configuration, breaking it into a list of configs, sending that list with `send_configs` and then joining the
   responses into a single `Response` object... or of course you can just send a single config line with it too!
+- Add better handling/logging for `SystemSSH` transport when key exchange cannot be negotiated
 
 # 2020.05.09
 - Add underscores to EOS config prompt matching